Disabling one warning so that we can use double hashtag __VA_ARGS__ extension.
Need to check this works properly in Windows build.

Merge branch 'master' into cmm_mkp

+++ include/aws/cryptosdk/cipher.h
@@ -29,7 +29,7 @@ struct aws_cryptosdk_alg_properties {
      */
     const struct aws_cryptosdk_alg_impl *impl;
 
-    int data_key_len, content_key_len, iv_len, tag_len;
+    size_t data_key_len, content_key_len, iv_len, tag_len;
 
     enum aws_cryptosdk_alg_id alg_id;
 };
